Add files via upload
To implement techniques of image processing to enhance 
the pattern of palatal rugae.Theuniqueness of palatal rugae 
pattern can be utilized similar to fingerprints and when 
combined with other methods, it canhelp in the 
identification of a person.Thus these unique characteristics 
of the palatal rugae can be studied and can be enhanced a 
using Image processing and can be used as identity for the 
each individual. By using the data samples of palates of 
different individual, uniqueness of the identity can be 
recognized. The image processing techniques can be applied
to these data samples to get the desired result.